refactor to pass back the valhalla error http code
