+++ src/bindings/node/node.cc
@@ -12,8 +12,7 @@
 #include <sstream>
 #include <string>
 
-#include "src/worker.cc"
-#include "valhalla/exception.h"
+#include "valhalla/worker.h"
 #include "valhalla/tyr/actor.h"
 #include "valhalla/midgard/logging.h"
 #include "valhalla/midgard/util.h"
@@ -46,7 +45,7 @@ public:
     try {
       response = func(actor, request);
     } catch (const valhalla::valhalla_exception_t& e) {
-      auto http_code = ERROR_TO_STATUS.find(e.code)->second;
+      auto http_code = e.http_code;
       rapidjson::StringBuffer err_message;
       rapidjson::Writer<rapidjson::StringBuffer> writer(err_message);
 
@@ -56,7 +55,7 @@ public:
       writer.Key("http_code");
       writer.Uint(http_code);
       writer.Key("message");
-      writer.String(e.message);
+      writer.String(e.message.c_str());
       writer.EndObject();
       throw std::runtime_error(err_message.GetString());
     } catch (const std::exception& e) { throw std::runtime_error(e.what()); }
