+++ src/tyr/route_serializer.cc
@@ -11,7 +11,6 @@
 #include "baldr/json.h"
 #include "baldr/rapidjson_utils.h"
 #include "baldr/turn.h"
-#include "exception.h"
 #include "midgard/aabb2.h"
 #include "midgard/encoded.h"
 #include "midgard/logging.h"
