+++ src/worker.cc
@@ -263,6 +263,7 @@ const std::unordered_map<unsigned, std::string> OSRM_ERRORS_CODES{
 
     {599, R"({"code":"InvalidUrl","message":"URL string is invalid."})"}};
 
+
 rapidjson::Document from_string(const std::string& json, const valhalla_exception_t& e) {
   rapidjson::Document d;
   d.Parse(json.c_str());
@@ -556,6 +557,17 @@ void from_json(rapidjson::Document& doc, odin::DirectionsOptions& options) {
 
 namespace valhalla {
 
+valhalla_exception_t::valhalla_exception_t(unsigned code, const boost::optional<std::string>& extra)
+      : std::runtime_error(""), code(code), extra(extra) {
+  auto code_iter = error_codes.find(code);
+  message = (code_iter == error_codes.cend() ? "" : code_iter->second);
+  message += (extra ? ":" + *extra : "");
+  auto http_code_iter = ERROR_TO_STATUS.find(code);
+  http_code = (http_code_iter == ERROR_TO_STATUS.cend() ? 0 : http_code_iter->second);
+  auto http_message_iter = HTTP_STATUS_CODES.find(http_code);
+  http_message = (http_message_iter == HTTP_STATUS_CODES.cend() ? "" : http_message_iter->second);
+}
+
 valhalla_request_t::valhalla_request_t() {
   document.SetObject();
 }
@@ -644,8 +656,6 @@ worker_t::result_t jsonify_error(const valhalla_exception_t& exception,
                                  http_request_info_t& request_info,
                                  const valhalla_request_t& request) {
   // get the http status
-  auto status = ERROR_TO_STATUS.find(exception.code)->second;
-  auto message = HTTP_STATUS_CODES.find(status)->second;
   std::stringstream body;
 
   // overwrite with osrm error response
@@ -660,8 +670,8 @@ worker_t::result_t jsonify_error(const valhalla_exception_t& exception,
   else {
     // build up the json map
     auto json_error = baldr::json::map({});
-    json_error->emplace("status", message);
-    json_error->emplace("status_code", static_cast<uint64_t>(status));
+    json_error->emplace("status", exception.http_message);
+    json_error->emplace("status_code", static_cast<uint64_t>(exception.http_code));
     json_error->emplace("error", std::string(exception.message));
     json_error->emplace("error_code", static_cast<uint64_t>(exception.code));
     body << (request.options.has_jsonp() ? request.options.jsonp() + "(" : "") << *json_error
@@ -669,7 +679,7 @@ worker_t::result_t jsonify_error(const valhalla_exception_t& exception,
   }
 
   worker_t::result_t result{false};
-  http_response_t response(status, message, body.str(),
+  http_response_t response(exception.http_code, exception.http_message, body.str(),
                            headers_t{CORS, request.options.has_jsonp() ? JS_MIME : JSON_MIME});
   response.from_info(request_info);
   result.messages.emplace_back(response.to_string());
